faster-rcnn/requirements.txt
visdom
torchvision
scikit-image
tqdm
fire
pprint
matplotlib
ipdb
cython
torchnet
cupy-cuda101
pillow==6.1